refactor(agent): extract prompt input wrappers
